bridge/requirements.txt
fastapi
uvicorn
playwright
google-generativeai
python-dotenv
pillow
python-multipart
